Moved camelcase util to under url namespace
